Force caller to pop aggregate return value param
Normally the callee is responsible for popping the
hidden parameter in a function that returns a non-
scalar value.  However, the gladiator bot's
gladi386.so binary appears to have been built such
that the hidden parameter is popped by the caller.

This change adds a special gcc attribute to the
one affected function in the open-source main game
library that has a non-scalar return type and is
called from the gladi386.so code:
BotLibImport_Trace().  This ensures the calling
conventions match between the gamei386.so and the
gladi386.so binary in all cases.  Note this
attribute shouldn't alter the existing behavior on
windows, since the MS ABI for x86_32 already uses
a caller-pop convention for aggregate return
value pointers.